plugins/autopara/scripts/pyproject.toml
[project]
name = "autopara-scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
